fix ut

+++ be/test/storage/rowset_column_update_state_test.cpp
@@ -223,7 +223,7 @@ static uint32_t find_upt_row_id(const ColumnPartialUpdateState& state, uint64_t 
     std::map<uint64_t, uint32_t> m;
     for (const auto& each_rss : state.rss_rowid_to_update_rowid) {
         for (const auto& each : each_rss.second) {
-            m[each_rss.first << 32 | each.first] = each.second;
+            m[(uint64_t)each_rss.first << 32 | (uint64_t)each.first] = each.second;
         }
     }
     return m[src_rss_id];
